move log indent decorator to cfclogger.py
